requirements.txt
numpy~=1.26.4
rasterio~=1.3.9
shapely~=2.0.3
pillow~=10.3.0
pandas~=2.2.2
geopandas~=0.14.3
scipy~=1.13.0
scikit-image~=0.23.1
tqdm~=4.66.2
scikit-learn~=1.4.2
affine~=2.4.0
pyproj~=3.6.1
shap~=0.46.0